vocalQ--Outbound-sivajanya_be/backend/requirements.txt
fastapi
uvicorn
supabase
pydantic
pydantic-settings
python-dotenv
websockets
httpx
openai
pandas
qdrant-client
python-dateutil
PyPDF2
python-multipart
python-docx
twilio
torch
numpy
scipy
